Fix lingering sign compare warnings in exposed by "ninja check-mlir".
--

PiperOrigin-RevId: 248050178

+++ examples/Linalg/Linalg1/lib/ConvertToLLVMDialect.cpp
@@ -210,7 +210,7 @@ public:
     // dimensions, extracts the size from the memref descriptor.
     auto memrefSize = [int64Ty, pos, i64cst](MemRefType type, Value *memref,
                                              int dim) -> Value * {
-      assert(dim < type.getRank());
+      assert(static_cast<size_t>(dim) < type.getRank());
       if (type.getShape()[dim] != -1) {
         return i64cst(type.getShape()[dim]);
       }
